Add files via upload
This is the first project in the Data Analyst Nanodegree. I wrote a SQL query to extract data provided by Udacity, and saved the data as CSV file. Then I used Excel as analyze and visualization tool to create moving average line chart. Observations and conclusions are made in the report.